--- samples/snippets/csharp/concepts/linq/how-to-group-query-results_4.cs
@@ -1,6 +1,0 @@
-﻿        //Helper method, used in GroupByRange.
-        protected static int GetPercentile(Student s)
-        {
-            double avg = s.ExamScores.Average();
-            return avg > 0 ? (int)avg / 10 : 0;
-        }
